requirements.txt
pandas==1.3.4
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-auth==2.3.2
google-api-python-client==2.28.0
google-api-core==2.2.0